- Melhorias e separação de código.
